server/skin-predictor/requirements.txt
flask
tensorflow
pillow
flask-cors
gunicorn

